requirements.txt
# 뉴스 수집 파이프라인 의존성
pandas>=2.0.0
feedparser>=6.0.10
python-dotenv>=1.0.0
requests>=2.31.0

